Refactor and remove EditTimeLog.js and imports, add loading state to NewTimeLog and isEditable prop to control form logic
