--- src/components/WorkOrder/EditTimeLog.js
@@ -1,257 +1,0 @@
-import React, { Component } from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faClock } from "@fortawesome/free-regular-svg-icons";
-import { faSpinner } from "@fortawesome/free-solid-svg-icons";
-import Select from "react-select";
-import { Redirect } from "react-router-dom";
-
-import Header from "../Shared/Header";
-import { ErrorMessage } from "./Alerts";
-import EditTimeLogDateTimeFields from "./EditTimeLogDateTimeFields";
-import SubmitButton from "../Form/SubmitButton";
-import api from "../../queries/api";
-import { FIELDS } from "./formConfig";
-
-import "react-datepicker/dist/react-datepicker.css";
-
-class EditTimeLog extends Component {
-  constructor(props) {
-    super(props);
-
-    this.workOrderId = this.props.match.params.workOrderId;
-
-    this.state = {
-      technicianOptions: [],
-      vehicleOptions: [],
-      timeLogData: [],
-      timeLogToEdit: null,
-      updatedFormData: { field_1424: this.workOrderId },
-      isFormDisabled: false,
-    };
-  }
-
-  submitForm = e => {
-    e.preventDefault();
-    this.setState({ errors: [], isSubmitting: true });
-
-    console.log("submitting: ", this.state.updatedFormData);
-
-    console.log(this.workOrderId);
-    api
-      .workOrder()
-      .editTimeLog(
-        this.props.match.params.timeLogId,
-        this.state.updatedFormData
-      )
-      .then(res => {
-        console.log(res);
-        this.setState({
-          isSubmitting: false,
-          isSubmitted: true,
-          successfulResponseData: res.data.record,
-        });
-      })
-      .catch(error => {
-        console.log(error.response.data.errors);
-        this.setState({
-          errors: error.response.data.errors,
-          isSubmitting: false,
-        });
-      });
-  };
-
-  handleChange = e => {
-    let updatedFormData = this.state.updatedFormData;
-    updatedFormData[e.target.name] = e.target.value;
-    const updatedAllData = Object.assign(
-      {},
-      this.state.rawData,
-      updatedFormData
-    );
-    updatedAllData[`${e.target.name}_raw`] = "";
-    this.setState({ updatedFormData, rawData: updatedAllData });
-  };
-
-  handleReactSelectChange = (fieldId, selected) => {
-    let updatedData = selected ? selected.value : "";
-
-    // create object of updated data
-    let updatedFormData = Object.assign({}, this.state.updatedFormData, {
-      [fieldId]: updatedData,
-    });
-
-    // merge updated data into all data
-    const updatedAllData = Object.assign(
-      {},
-      this.state.rawData,
-      updatedFormData
-    );
-
-    delete updatedAllData[`${fieldId}_raw`];
-
-    this.setState({ updatedFormData, rawData: updatedAllData });
-  };
-
-  handleFormDisable = boolean => {
-    this.setState({ isFormDisabled: boolean });
-  };
-
-  handleReactMultiSelectChange = (name, values) => {
-    // React-Select sends the event as the updated selected values.
-    // https://github.com/JedWatson/react-select/issues/1631
-    let data = {};
-    data[name] = values.map(item => item.value);
-
-    // create object of updated data
-    let updatedFormData = Object.assign({}, this.state.updatedFormData, data);
-
-    // merge updated data into all data
-    const updatedAllData = Object.assign(
-      {},
-      this.state.rawData,
-      updatedFormData
-    );
-
-    this.setState({ updatedFormData, rawData: updatedAllData });
-  };
-
-  getTechnicianOptions = () => {
-    api
-      .workOrder()
-      .getTimeLogTechnicianOptions()
-      .then(res => {
-        const technicianOptions = res.data.records.map(item => {
-          return { label: item.identifier, value: item.id };
-        });
-        this.setState({ technicianOptions });
-      });
-  };
-
-  getVehicleOptions = () => {
-    api
-      .workOrder()
-      .getVehicleOptions()
-      .then(res => {
-        const vehicleOptions = res.data.records.map(item => {
-          return { label: item.identifier, value: item.id };
-        });
-        this.setState({ vehicleOptions });
-      });
-  };
-
-  setTimeLogToEdit = () => {
-    const timeLogToEdit = this.state.timeLogData.filter(
-      record => record.id === this.props.match.params.timeLogId
-    )[0];
-    this.setState({ timeLogToEdit });
-  };
-
-  requestTimeLogs = id => {
-    api
-      .workOrder()
-      .getTimeLogs(id)
-      .then(res =>
-        this.setState({ timeLogData: res.data.records }, () => {
-          this.setTimeLogToEdit();
-        })
-      );
-  };
-
-  handleDateTimeFieldChange = (fieldId, fieldData) => {
-    let formData = this.state.updatedFormData;
-
-    formData[fieldId] = fieldData;
-    this.setState({ updatedFormData: formData });
-  };
-
-  getEditFieldDefaultValues = vehiclesArray => {
-    vehiclesArray.forEach(record => {
-      record["label"] = record.identifier;
-      record["value"] = record.id;
-    });
-    return vehiclesArray;
-  };
-
-  componentDidMount() {
-    this.getTechnicianOptions();
-    this.getVehicleOptions();
-    this.requestTimeLogs(this.workOrderId);
-  }
-
-  render() {
-    if (!!this.state.successfulResponseData) {
-      console.log(this.state.successfulResponseData);
-      return <Redirect to={`/work-orders/${this.workOrderId}`} />;
-    }
-
-    const timeLogToEdit = this.state.timeLogToEdit;
-
-    return this.state.timeLogToEdit ? (
-      <div>
-        <Header icon={faClock} title="Edit Time Log" />
-
-        {this.state.errors &&
-          this.state.errors.map(error => (
-            <ErrorMessage error={error} key={error.field} />
-          ))}
-
-        <form onSubmit={this.submitForm}>
-          <div className="form-group">
-            <label htmlFor={FIELDS.TIMELOG.TECHNICIANS}>Technician(s)</label>
-            <Select
-              className="basic-multi-select"
-              classNamePrefix="select"
-              isMulti
-              defaultValue={this.getEditFieldDefaultValues(
-                timeLogToEdit[FIELDS.TIMELOG.EDIT_TECHNICIANS]
-              )}
-              id={FIELDS.TIMELOG.TECHNICIANS}
-              name={FIELDS.TIMELOG.TECHNICIANS}
-              aria-describedby={`${FIELDS.TIMELOG.TECHNICIANS}-text`}
-              placeholder={"Leave blank for yourself"}
-              options={this.state.technicianOptions}
-              onChange={this.handleReactMultiSelectChange.bind(
-                this,
-                FIELDS.TIMELOG.TECHNICIANS
-              )}
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor={FIELDS.TIMELOG.VEHICLES}>Vehicles(s)</label>
-            <Select
-              className="basic-multi-select"
-              classNamePrefix="select"
-              isMulti
-              defaultValue={this.getEditFieldDefaultValues(
-                timeLogToEdit[FIELDS.TIMELOG.EDIT_VEHICLES]
-              )}
-              id={FIELDS.TIMELOG.VEHICLES}
-              name={FIELDS.TIMELOG.VEHICLES}
-              options={this.state.vehicleOptions}
-              onChange={this.handleReactMultiSelectChange.bind(
-                this,
-                FIELDS.TIMELOG.VEHICLES
-              )}
-            />
-          </div>
-          <EditTimeLogDateTimeFields
-            data={this.state.updatedFormData}
-            handleTimeChange={this.handleDateTimeFieldChange}
-            handleFormDisable={this.handleFormDisable}
-            isFormDisabled={this.state.isFormDisabled}
-            timeLogToEdit={timeLogToEdit}
-          />
-          <SubmitButton
-            text="Edit Log Entry"
-            isSubmitting={this.state.isSubmitting}
-            isFormDisabled={this.state.isFormDisabled}
-          />
-        </form>
-      </div>
-    ) : (
-      <FontAwesomeIcon icon={faSpinner} className="atd-spinner" size="2x" />
-    );
-  }
-}
-
-export default EditTimeLog;
